Log # of unused Zs
Since fixed point finder is parallelized, it may consume different
number of Zs depending on how fast one CPU finishes the computation
comparing to others.  It makes us unable to reproduce simulations
exactly.  If we record the number of unused Zs (and the seed), it is in
principle possible to do the same simulation by tracing this number.

(I doubt we really need these numbers and it's just me being paranoid,
but there is not much harm in saving those numbers.)