Support array value for update clip
When the clipping min and max values are arrays (list) and Theano
device is set to GPU (by THEANO_FLAGS=device=cuda,floatX=float32), the
update does not compile.  The min and max values have to be explicitly
converged to supported dtype.

It can be verified that this change is necessary by::

  THEANO_FLAGS=device=cuda,floatX=float32,force_device=True misc/with-env pytest -k 'test_single_g_step_with_load_config_slowtest[args3-config3]'

Note that "misc/with-env pytest" has to be used instead of
"misc/pytest", since this script adds mode=FAST_COMPILE.  It seems
Theano does not like device=cuda with mode=FAST_COMPILE.  (Using
optimizer=fast_compile instead of mode=FAST_COMPILE seems to work.
Maybe I should use it in misc/pytest.)

This pytest command reports failures of the two test just added if the
two np.array lines added are removed.